fix: clean only if needed

+++ builder/builder.py
@@ -438,8 +438,9 @@ class Builder:
 		if self.user_user_config:
 			self.user_user_config = utils.pathExpand(self.user_user_config)
 
-		if os.path.exists(self.dir_build):
-			shutil.rmtree(self.dir_build)
+		if self.build_clean:
+			if os.path.exists(self.dir_build):
+				shutil.rmtree(self.dir_build)
 
 
 	def config(self):
